# Improve the Tokenizer performance