@tgalopin 2018-11-05
I just added an additional commit to inline the parsing of tags too (critical path as well and easy to inline without duplication). The new benchmarks using the same script are:

*Before*:

```
Total time: 2104ms
Time per iteration: 21.04ms
```

*After*:

```
Total time: 1626ms
Time per iteration: 16.26ms
```