@tgalopin 2018-11-06
I just added another improvement from the idea of @mundschenk-at:

```
Total time: 1580ms
Time per iteration: 15.8ms
```

Given how impactful each `current` call removal is, I will try in another PR to reduce the number of calls even more.